Extract methods for data manipulation to a separate class
